GPP-545 statistics api update

+++ geoss-proxy/src/main/java/com/eversis/esa/geoss/proxy/domain/MostPopular.java
@@ -20,6 +20,6 @@ public class MostPopular {
     /**
      * The Doc count.
      */
-    int docCount;
+    long docCount;
 
 }
