+++ geoss-proxy/src/main/java/com/eversis/esa/geoss/proxy/domain/Search.java
@@ -25,6 +25,6 @@ public class Search {
     /**
      * The Doc count.
      */
-    int docCount;
+    long docCount;
 
 }
